requirements.txt
streamlit
ultralytics
numpy
opencv-python
Pillow
fastapi
uvicorn
requests
python-multipart
python-dotenv